Literalize `$text_domain` parameter in all gettext functions.
When internationalizing user interface text, the $text_domain parameter
must not be a variable. This is because the tools that parse plugin
source code in order to map the strings to the correct text domains do
not actually use PHP interpreters and can therefor not interpolate the
variable values to the correct string.

See also this warning:
https://developer.wordpress.org/plugins/internationalization/how-to-internationalize-your-plugin/#text-domains